+++ src/jvmMain/kotlin/net/longbowxxx/playground/logger/ChatLogger.kt
@@ -19,10 +19,21 @@ class ChatLogger(
         const val CHAT_LOG_DIR = "$LOG_DIR/chat"
     }
 
+    private var headerWrote = false
+
+    private suspend fun logHeaderIfNeeds() {
+        if (!headerWrote) {
+            headerWrote = true
+            writeLog {
+                write("Playground Chat Log $dateTimeStr\n")
+                write("===\n")
+            }
+        }
+    }
+
     suspend fun logRequest(request: OpenAiChatRequest) {
+        logHeaderIfNeeds()
         writeLog {
-            write("Playground Chat Log $dateTimeStr\n")
-            write("===\n")
             write("# Request Parameters\n")
             write("```json\n")
             val json = encodeJson.encodeToString(request.extractSettings())
@@ -33,10 +44,18 @@ class ChatLogger(
     }
 
     suspend fun logMessages(messages: List<OpenAiChatMessage>) {
+        logHeaderIfNeeds()
         writeLog {
             messages.forEach { message ->
                 write("# ${message.role.name}\n")
-                write("${message.content}\n")
+                if (message.content != null) {
+                    write("${message.content}\n")
+                }
+                if (message.functionCall != null) {
+                    write("## function call  \n")
+                    write("${message.functionCall?.name}  \n")
+                    write("```json\n${message.functionCall?.arguments}\n```\n")
+                }
                 appendHorizontalLine()
             }
         }