+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 PlaygroundL Change Log
 ===
 
+## [1.5.0]
+### Added
+- Add plugin system for function
+  - Add create image plugin
+  - Add save file plugin
+- Add PaLM2 discuss service (English Only)
+- Add github actions
+
 ## [1.4.0]
 ### Added
 - Add chat function call